# apps/api/openapi/agent-tools.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pi Finance Agent HTTP contract",
    "version": "1.0.0"
  },
  "paths": {
...
    "/cards/recurring": {
      "get": {
        "operationId": "list_recurring_purchases",
        "summary": "List recurring credit-card purchases",
        "x-pi-tool": {
          "name": "list_recurring_purchases",
          "label": "list_recurring_purchases",
          "description": "List recurring credit-card purchases",
          "method": "GET",
          "path": "/cards/recurring",
          "idempotency": false,
          "shadow": false,
          "result": null
        },
        "parameters": [
          {
            "name": "householdId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "x-pi-context": true
          },
          {
            "name": "accountId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "paused",
                "cancelled"
              ]
            }
          }
        ]
      },
      "post": {
        "operationId": "create_recurring_purchase",
        "summary": "HTTP contract for create_recurring_purchase",
        "x-pi-tool": {
          "name": "create_recurring_purchase",
          "label": "create_recurring_purchase",
          "description": "HTTP contract for create_recurring_purchase",
          "method": "POST",
          "path": "/cards/recurring",
          "idempotency": true,
          "shadow": false,
          "result": null
        },
        "parameters": [
          {
            "name": "householdId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          },
          {
            "name": "idempotencyKey",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "accountId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "description": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 240
                  },
                  "amountCents": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "frequency": {
                    "type": "string",
                    "enum": [
                      "monthly",
                      "quarterly",
                      "yearly"
                    ]
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "categoryId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "accountId",
                  "description",
                  "amountCents",
                  "frequency",
                  "startDate"
                ]
              }
            }
          }
        }
      }
    },
...
    "/cards": {
      "post": {
        "operationId": "create_credit_card_account",
        "summary": "HTTP contract for create_credit_card_account",
        "x-pi-tool": {
          "name": "create_credit_card_account",
          "label": "create_credit_card_account",
          "description": "HTTP contract for create_credit_card_account",
          "method": "POST",
          "path": "/cards",
          "idempotency": true,
          "shadow": false,
          "result": null
        },
        "parameters": [
          {
            "name": "householdId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          },
          {
            "name": "idempotencyKey",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "creditLimitCents": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "closingDay": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 31
                  },
                  "dueDay": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 31
                  }
                },
                "required": [
                  "name",
                  "creditLimitCents",
                  "closingDay",
                  "dueDay"
                ]
              }
            }
          }
        }
      }
    },
    "/cards/purchases": {
      "post": {
        "operationId": "create_card_purchase",
        "summary": "HTTP contract for create_card_purchase",
        "x-pi-tool": {
          "name": "create_card_purchase",
          "label": "create_card_purchase",
          "description": "HTTP contract for create_card_purchase",
          "method": "POST",
          "path": "/cards/purchases",
          "idempotency": true,
          "shadow": false,
          "result": null
        },
        "parameters": [
          {
            "name": "householdId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          },
          {
            "name": "idempotencyKey",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "accountId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "description": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 240
                  },
                  "amountCents": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "date": {
                    "type": "string",
                    "format": "date",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "categoryId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "subcategoryId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "installmentsTotal": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 48
                  },
                  "installmentNumber": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 48
                  }
                },
                "required": [
                  "accountId",
                  "description",
                  "amountCents",
                  "date"
                ]
              }
            }
          }
        }
      }
    },
    "/cards/installments": {
      "post": {
        "operationId": "create_card_installments",
        "summary": "HTTP contract for create_card_installments",
        "x-pi-tool": {
          "name": "create_card_installments",
          "label": "create_card_installments",
          "description": "HTTP contract for create_card_installments",
          "method": "POST",
          "path": "/cards/installments",
          "idempotency": true,
          "shadow": false,
          "result": null
        },
        "parameters": [
          {
            "name": "householdId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          },
          {
            "name": "idempotencyKey",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "accountId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "description": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 240
                  },
                  "totalAmountCents": {
                    "type": "integer",
                    "minimum": 100
                  },
                  "purchaseDate": {
                    "type": "string",
                    "format": "date",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "installmentsTotal": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 48
                  },
                  "categoryId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "subcategoryId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 2000
                  }
                },
                "required": [
                  "accountId",
                  "description",
                  "totalAmountCents",
                  "purchaseDate",
                  "installmentsTotal"
                ]
              }
            }
          }
        }
      }
    },
    "/cards/statements/{statementId}/pay": {
      "post": {
        "operationId": "pay_statement",
        "summary": "HTTP contract for pay_statement",
        "x-pi-tool": {
          "name": "pay_statement",
          "label": "pay_statement",
          "description": "HTTP contract for pay_statement",
          "method": "POST",
          "path": "/cards/statements/{statementId}/pay",
          "idempotency": true,
          "shadow": false,
          "result": null
        },
        "parameters": [
          {
            "name": "householdId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          },
          {
            "name": "idempotencyKey",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          },
          {
            "name": "statementId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amountCents": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "fromAccountId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "amountCents",
                  "fromAccountId"
                ]
              }
            }
          }
        }
      }
    },
    "/cards/statements": {
      "get": {
        "operationId": "list_statements",
        "summary": "HTTP contract for list_statements",
        "x-pi-tool": {
          "name": "list_statements",
          "label": "list_statements",
          "description": "HTTP contract for list_statements",
          "method": "GET",
          "path": "/cards/statements",
          "idempotency": false,
          "shadow": true,
          "result": null
        },
        "parameters": [
          {
            "name": "householdId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          },
          {
            "name": "accountId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "closed",
                "paid",
                "partial",
                "overdue",
                "cancelled"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          }
        ]
      }
    },
    "/cards/statements/{statementId}": {
      "get": {
        "operationId": "get_statement_details",
        "summary": "HTTP contract for get_statement_details",
        "x-pi-tool": {
          "name": "get_statement_details",
          "label": "get_statement_details",
          "description": "HTTP contract for get_statement_details",
          "method": "GET",
          "path": "/cards/statements/{statementId}",
          "idempotency": false,
          "shadow": true,
          "result": null
        },
        "parameters": [
          {
            "name": "householdId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          },
          {
            "name": "statementId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
...
  }
}